+++ tests/ContactsTest/Contacts.h
@@ -35,6 +35,14 @@
    Reads the status of contact sensors
 */
 class ContactSensors {
+private:
+  const uint8_t _frontSensorPin;
+  const uint8_t _rearSensorPin;
+  boolean _frontClear;
+  boolean _rearClear;
+  void *_context;
+  void (*_onChanged)(void *, ContactSensors &);
+
 public:
   /*
        Creates the contact sensors
@@ -65,11 +73,13 @@ public:
     return _rearClear;
   }
 
-private:
-  const uint8_t _frontSensorPin;
-  const uint8_t _rearSensorPin;
-  boolean _frontClear;
-  boolean _rearClear;
+  /**
+       Sets callback on reached
+    */
+  void onChanged(void (*callback)(void *context, ContactSensors &sensors), void *context = NULL) {
+    _onChanged = callback;
+    _context = context;
+  }
 };
 
 #endif