+++ tests/MotorsTest/pins.h
@@ -45,8 +45,8 @@
   22 - SCL
   23 - TRIGGER_PIN
   25 - STATUS_LED_PIN
-  26 - 
-  27 - 
+  26 - FRONT_LIDAR_PIN
+  27 - REAR_LIDAR_PIN
   32 - LEFT_PIN
   33 - RIGHT_PIN
   34 - ECHO_PIN - input only
@@ -57,32 +57,34 @@
 
 // The ADC pins
 
-#define VOLTAGE_PIN         39
+#define VOLTAGE_PIN 39
 
 // The digital input pins
 
-#define LEFT_PIN            32
-#define RIGHT_PIN           33
-#define ECHO_PIN            34
-#define FRONT_CONTACTS_PIN  35
-#define REAR_CONTACTS_PIN   36
+#define LEFT_PIN 32
+#define RIGHT_PIN 33
+#define ECHO_PIN 34
+#define FRONT_CONTACTS_PIN 35
+#define REAR_CONTACTS_PIN 36
 
 // The PWM output pins
 
-#define SERVO_PIN       13
-#define LEFT_FORW_PIN   16
-#define LEFT_BACK_PIN   17
-#define RIGHT_FORW_PIN  18
-#define RIGHT_BACK_PIN  19
+#define SERVO_PIN 13
+#define LEFT_FORW_PIN 16
+#define LEFT_BACK_PIN 17
+#define RIGHT_FORW_PIN 18
+#define RIGHT_BACK_PIN 19
 
 // IC2 pins
 
-#define SDA_PIN   21
-#define SCL_PIN   22
+#define SDA_PIN 21
+#define SCL_PIN 22
 
 // The digital output pins
 
-#define TRIGGER_PIN     23
-#define STATUS_LED_PIN  25
+#define TRIGGER_PIN 23
+#define STATUS_LED_PIN 25
+#define FRONT_LIDAR_PIN 26
+#define REAR_LIDAR_PIN 27
 
 #endif