+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- Issue [#117](https://github.com/m-marini/wheellino/issues/117): Add Lidar sensors
+
 ## [0.9.0] 2025-10-06
 
 ### Added